requirements.txt
JSON-log-formatter==0.4.0
psutil==5.8.0
pyaml==21.8.3
schedule==1.1.0
shiv==0.5.2
pytest==6.2.5
wheel
